Add proxy parser module
