+++ src/modules/proxy.ts
@@ -1,11 +1,22 @@
-// prefixRegex, _ := regexp.Compile(`(\d+\.){3}`)
-// suffixRegex, _ := regexp.Compile("(:.+){2,3}")
-// prefix := prefixRegex.FindString(ip)
-// suffix := suffixRegex.FindString(ip)
-// if prefix == "" {
-//     return [256]string{}, errors.New("improperly formatted IP address")
-// }
-// ipRange := [256]string{}
-// for i := 0; i < 256; i++ {
-//     ipRange[i] = fmt.Sprintf("%s%d%s", prefix, i, suffix)
-// }
+const parseProxy = (proxy: string): string[] => {
+    const prefix: RegExpMatchArray | null = proxy.match(/(?:\d+\.){3}/);
+    const suffix: RegExpMatchArray | null = proxy.match(/(?::.+){2,3}/);
+
+    if (prefix === null) {
+        throw new Error("Improperly formatted IP address (prefix)");
+    }
+    if (suffix === null) {
+        throw new Error("Improperly formatted IP address (suffix)");
+    }
+
+    const prefixString: string = prefix[0];
+    const suffixString: string = suffix[0];
+    const ipRange: string[] = [];
+    for (let i = 0; i < 256; i++) {
+        ipRange.push(`${prefixString}${i}${suffixString}`);
+    }
+
+    return ipRange;
+}
+
+export {parseProxy};
