+++ src/modules/variants.ts
@@ -1,22 +1,7 @@
 import axios, { AxiosResponse } from "axios";
 import { ProductVariants } from "../types/product";
 import { load } from "cheerio";
-
-const headers = {
-	accept: "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
-	"accept-encoding": "gzip, deflate",
-	"accept-language": "en-US,en;q=0.9",
-	dnt: "1",
-	"sec-ch-ua": '"Google Chrome";v="105", "Not)A;Brand";v="8", "Chromium";v="105"',
-	"sec-ch-ua-mobile": "?0",
-	"sec-ch-ua-platform": '"Windows"',
-	"sec-fetch-dest": "document",
-	"sec-fetch-mode": "navigate",
-	"sec-fetch-site": "none",
-	"sec-fetch-user": "?1",
-	"upgrade-insecure-requests": "1",
-	"user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/105.0.0.0 Safari/537.36",
-};
+import { headers } from "../common/requests";
 
 const getVariants = async (url: string) => {
 	let resp: AxiosResponse;
